requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.17.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
environs==9.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
ibm-cloud-sdk-core==3.10.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-watson-openscale==3.0.14
ibm_watsonx_ai==1.1.8
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lomond==0.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
milvus-lite==2.4.10
msal==1.31.0
msal-extensions==1.2.0
narwhals==1.8.1
numpy==1.26.4
openai==1.45.1
packaging==24.1
pandas==2.1.4
pillow==10.4.0
portalocker==2.10.1
protobuf==5.28.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymilvus==2.4.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.2
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.38.0
tabulate==0.9.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
zipp==3.20.2
